playing with Bayesian imputation using simulated data
Used Binomial and Poisson as an easy starting point to model the unobserved ICS stars, but multinomial or ordered_logit would be a better solution. Would probably switch to Stan for a final model, but started with JAGS because Stan does not allow latent discrete parameters (e.g. Poisson, Binomial).